Heavily refactored the Model, the main reason for this drastic change is due to: - Since different gizmos such as SquareBumper & Ball uses different precision methods (for board placement, even in Build Mode), it is extremely confusing for AGizmoComponent (parent class of all Gizmos) to use Int for X & Y. Not only that, it is very easy to use the wrong X/Y methods for the extra precision Gizmos, thus further complicating coding process. - Similar to the reason above, bmWidth & bmHeight does not exist for Moving Gizmos (e.g. Ball), so having those fields in the AGizmoComponent is incorrect.
As a consequence of this, some methods (particularly in the MainEngine) that deals with board placement will need to be altered:
Different block of code is needed to deal with the different precision X & Y methods, even if it is simply due to different return types etc.

In summary:
This commit causes more code writing, but in return much easier-to-understand & maintainable code, as well as extra method misuse protection due to the X/Y methods being completely different.